CPlay Express; speaker and talkie functions

+++ utility/Adafruit_CPlay_Speaker.h
@@ -12,7 +12,7 @@ class Adafruit_CPlay_Speaker {
   void begin(void),
        end(void),
        set(uint8_t value),
-       playSound(const uint8_t *data, uint16_t length, uint16_t sampleRate),
+       playSound(const uint8_t *data, uint32_t length, uint16_t sampleRate),
        say(const uint8_t *address);
 
  private:
